Add files via upload
In this project I have analyzed data obtained from a series of measurements of a freefalling object. I've run the data through Chauvenet's Criterion, in order to filter out the outliers. Than I performed chi-squared test to test the hypothesis that the gathered data follows Gaussian distribution. Lastly, I've used polynomial interpolation and linear regression, along with the error analysis methods, to prove that the measurements agree closely with the referent value for the gravity acceleration.