Add files via upload
In this project I have loaded the various tables from the Rebrickable website and perform analysis on the data. Rebrickable is a fantastic website that is used to show which LEGO sets we can build from the sets and parts we already own. They will even show which parts we are missing to complete the new set and suggest sets that we could purchase to get those parts. 

Using the pandas library I have manipulated the dataset to answer various questions about, and perform an analysis of the dataset.